правки

+++ webserver_freertos/example/src/stand.h
@@ -31,8 +31,9 @@ typedef struct{
 
 typedef struct{
 	//int32_t posImp;
-	uint32_t div;
+	uint16_t div;
 	uint8_t dir;
+
 	//uint32_t time;
 	//uint32_t num;
 } TMoveCmd;
